backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
channels>=4.0
channels-redis>=4.2
yt-dlp>=2024.0
openai>=1.0
anthropic>=0.18
python-dotenv>=1.0
gunicorn>=21.0
uvicorn>=0.27
requests
daphne